refactor: move component in the page/dataverse folder
