+++ src/page/dataverse/component/generalMetadata/generalMetadata.scss
