+++ src/page/dataverse/component/generalMetadata/i18n/generalMetadata_fr.json
