+++ src/ui/page/dataverse/service/service.tsx
@@ -4,8 +4,8 @@ import type { Option } from 'fp-ts/Option'
 import { match, none } from 'fp-ts/Option'
 import { getResourceDetails } from '@/ui/page/dataverse/dataverse'
 import type { DataverseItemDetails } from '@/ui/page/dataverse/dataverse'
-import type { GeneralMetadata } from '@/component/generalMetadata/generalMetadata'
-import { GeneralMetadataList } from '@/component/generalMetadata/generalMetadata'
+import type { GeneralMetadata } from '@/page/dataverse/component/generalMetadata/generalMetadata'
+import { GeneralMetadataList } from '@/page/dataverse/component/generalMetadata/generalMetadata'
 
 const serviceGeneralMetadata: GeneralMetadata[] = [
   {
